refactor: Extract ProfileForm logic to custom hook and centralize schema
- Create useProfileForm hook with all form logic, state management, and API calls
- Move profileSchema to types file for centralized validation
- Infer ProfileFormData type from schema using z.infer
- Simplify ProfileForm component to focus on UI rendering
- Update profile-demo.ts to match new API signature
- Improve code organization and maintainability
